requirements.txt
dotenv
fastapi[standard]
langchain
langchain-google-genai
pymupdf 
pillow 
google-generativeai 
python-dotenv
fastapi
uvicorn
python-multipart
python-docx
